refactor: convert cost and token to float 32 precision

+++ ragaai_catalyst/tracers/tracer.py
@@ -15,7 +15,7 @@ from ragaai_catalyst.tracers.utils.langchain_tracer_extraction_logic import lang
 from ragaai_catalyst.tracers.upload_traces import UploadTraces
 import tempfile
 import json
-
+import numpy as np
 from opentelemetry.sdk import trace as trace_sdk
 from opentelemetry.sdk.trace.export import SimpleSpanProcessor
 from ragaai_catalyst.tracers.exporters.file_span_exporter import FileSpanExporter
@@ -297,13 +297,13 @@ class Tracer(AgenticTracing):
 
             # with open(filepath, 'r') as f:
             #     data = json.load(f)
-            additional_metadata["total_tokens"] = additional_metadata["tokens"]["total"]
+            additional_metadata["total_tokens"] = np.float32(additional_metadata["tokens"]["total"])
             del additional_metadata["tokens"]
             if "cost" in additional_metadata:
-                additional_metadata["total_cost"] = additional_metadata["cost"]["total_cost"]
+                additional_metadata["total_cost"] = np.float32(additional_metadata["cost"]["total_cost"])
                 del additional_metadata["cost"]
             else:
-                additional_metadata["total_cost"] = 0.0
+                additional_metadata["total_cost"] = np.float32(0.0)
             
             combined_metadata = user_detail['trace_user_detail']['metadata'].copy()
             combined_metadata.update(additional_metadata)
